refactor: modularize server logic
